source/Sources/frontend_backend/requirements.txt
flask
flask_pymongo
pymongo>=4.0.0
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv>=0.19.0
email-validator==2.0.0
requests>=2.25.0
mistralai>=0.0.7
beautifulsoup4
pandas
openpyxl
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
